vk: implement first version of vulkan backend
Graphics.Vulkan/ implements the backend API on Vulkan 1.3: dynamic rendering, synchronization2, scalar block layout, shader draw parameters, push descriptors, VK_EXT_memory_budget. Concretely:

- Automatic barriers. Each texture and buffer carries its layout and last access; every use emits the transition it needs and nothing when it doesn't. That's what reproduces D3D11's "bind it and it works".
- Rendering opens at the draw, not when targets are set. The caller binds targets before textures, and Vulkan can't transition an image inside a render pass — so the pass is deferred until everything is in the right layout.
- Renaming is real. An upload buffer holds one copy per frame in flight and rotates; Map(WriteDiscard) writes the copy this frame owns, so it never stalls and never corrupts a frame still on the GPU.
- Deferred destruction tied to frame fences; staging textures are buffers (what D3D11 actually uses them for, and every driver supports it); placeholder descriptors for bindings a draw leaves unset, because operators do that constantly.

The design change running it forced

Decision 10 said one descriptor set per shader stage. Testing against slangc showed that's unreachable: Slang derives the descriptor set from the HLSL register space, and none of TiXL's ~1000 shaders declare one — it would mean editing every shader. What does work, verified: offsetting each stage's bindings inside set 0 via the compiler's register shifts (-fvk-t-shift 232 0 → binding 232, set 0). So there's now one shared ShaderSlots table (stage stride 200; s/b/t/u at 0/16/32/160) that the compiler, the compat layer and both backends agree on, and SetBindings takes a ShaderStage instead of a set index.

+++ Graphics.Compat/DeviceContext.cs
@@ -306,7 +306,7 @@ public sealed class DeviceContext
         // Always sent, even when empty: a stage that bound something for the previous draw has to be
         // cleared, or its descriptors stay live.
         var count = stage.CollectBindings(_bindings);
-        Commands.SetBindings((int)stage.Stage, _bindings.AsSpan(0, count));
+        Commands.SetBindings(stage.Stage, _bindings.AsSpan(0, count));
     }
 
     /// <summary>
